creating services abstractions
