+++ server/TWS Foundation/TWS Customer/Features/Business/LocationsService.cs
@@ -1,6 +1,6 @@
-﻿using CSM_Foundation.Customer;
-using CSM_Foundation.Database.Entity.Models;
+﻿using CSM_Foundation.Database.Entity.Models;
 using CSM_Foundation.Database.Entity.Models.Output;
+using CSM_Foundation.Product;
 
 using TWS_Business;
 using TWS_Business.Depots;
@@ -40,7 +40,7 @@ public class LocationsService
 
         foreach (Location entity in Entities) {
             try {
-                Location attachedEntity = await _depot.Store(entity);
+                Location attachedEntity = await depot.Store(entity);
                 successes = [.. successes, attachedEntity];
             } catch (Exception excep) {
                 if (Sync) {
