+++ server/TWS Foundation/TWS Customer/Features/Business/SituationsService.cs
@@ -1,6 +1,6 @@
-﻿using CSM_Foundation.Customer;
-using CSM_Foundation.Database.Entity.Models;
+﻿using CSM_Foundation.Database.Entity.Models;
 using CSM_Foundation.Database.Entity.Models.Output;
+using CSM_Foundation.Product;
 
 using TWS_Business;
 using TWS_Business.Depots.Indicators;
@@ -40,7 +40,7 @@ public class SituationsService
 
         foreach (Situation entity in Entities) {
             try {
-                Situation attachedEntity = await _depot.Store(entity);
+                Situation attachedEntity = await depot.Store(entity);
                 successes = [.. successes, attachedEntity];
             } catch (Exception excep) {
                 if (Sync) {
