+++ server/TWS Foundation/TWS Customer/Features/Business/TrailersService.cs
@@ -1,6 +1,6 @@
-﻿using CSM_Foundation.Customer;
-using CSM_Foundation.Database.Entity.Models;
+﻿using CSM_Foundation.Database.Entity.Models;
 using CSM_Foundation.Database.Entity.Models.Output;
+using CSM_Foundation.Product;
 
 using TWS_Business;
 using TWS_Business.Entities.Trailers;
@@ -40,7 +40,7 @@ public class TrailersService
 
         foreach (Trailer_Common entity in Entities) {
             try {
-                Trailer_Common attachedEntity = await _depot.Store(entity);
+                Trailer_Common attachedEntity = await depot.Store(entity);
                 successes = [.. successes, attachedEntity];
             } catch (Exception excep) {
                 if (Sync) {
