+++ server/TWS Foundation/TWS Customer/Features/Business/TrucksService.cs
@@ -1,6 +1,6 @@
-﻿using CSM_Foundation.Customer;
-using CSM_Foundation.Database.Entity.Models;
+﻿using CSM_Foundation.Database.Entity.Models;
 using CSM_Foundation.Database.Entity.Models.Output;
+using CSM_Foundation.Product;
 
 using TWS_Business;
 using TWS_Business.Depots;
@@ -40,7 +40,7 @@ public class TrucksService
 
         foreach (Truck_Common entity in Entities) {
             try {
-                Truck_Common attachedEntity = await _depot.Store(entity);
+                Truck_Common attachedEntity = await depot.Store(entity);
                 successes = [.. successes, attachedEntity];
             } catch (Exception excep) {
                 if (Sync) {
