+++ server/TWS Foundation/TWS Customer/Features/Security/Profiles/IProfilesService.cs
@@ -1,13 +1,12 @@
-﻿using CSM_Foundation.Customer;
+﻿using CSM_Foundation.Product;
 
 using CSM_Security.Entities;
 
 namespace TWS_Customer.Features.Security.Profiles;
 
-
 /// <summary>
 ///     
 /// </summary>
-public interface IProfilesService 
+public interface IProfilesService
     : IService<Profile> {
 }
