+++ server/TWS Foundation/TWS Foundation/Controllers/Business/Vehicules/TrailersController.cs
@@ -11,29 +11,18 @@ using TWS_Foundation.Authentication;
 
 namespace TWS_Foundation.Controllers.Business.Vehicules;
 
-
-[ApiController, Route("[Controller]/[Action]")]
 [Feature("Trailers")]
 public class TrailersController
-    : ControllerBase {
+    : BEntityController<ITrailersService> {
 
-    /// <summary>
-    ///     Feature service dependency.
-    /// </summary>
-    readonly ITrailersService _service;
-
-    /// <summary>
-    ///     Creates a new instance.
-    /// </summary>
-    public TrailersController(ITrailersService service) {
-        _service = service;
+    public TrailersController(ITrailersService service) 
+        : base(service) {
     }
-
 
     [HttpPost, Action("View")]
     public async Task<IActionResult> View(ViewInput<Trailer_Common> input)
     => Ok(
-            await _service.View(
+            await service.View(
                     new QueryInput<Trailer_Common, ViewInput<Trailer_Common>> {
                         Parameters = input
                     }
